pyproject.toml
[tool.poetry]
name = "the-originals-test-task"
version = "0.1.0"
description = ""
authors = ["ArhipKokocan <hts1345hjyuhip228888@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
